Added all signUp requirements and added colour scheme
The changes to index.html are just redirecting the sign up button to the form. The forms close also is just a "Back" button it doesnt actually refer to the welcome page. I had trouble getting it to work as it still wanted me to fill out the required fields. Ill look into it later.